MAGECLOUD-3816: Eliminate Docker config.env file
