+++ src/Application.php
@@ -72,7 +72,7 @@ class Application extends \Symfony\Component\Console\Application
                 $this->container->create(Command\Wizard\IdealState::class),
                 $this->container->create(Command\Wizard\MasterSlave::class),
                 $this->container->create(Command\Docker\Build::class),
-                $this->container->create(Command\Docker\ConfigConvert::class),
+                $this->container->create(Command\Docker\GenerateDist::class),
                 $this->container->create(Command\CronKill::class),
                 $this->container->create(Command\ConfigShow::class),
             ]
